Pushed sync tag function up to the caller
